+++ NEWS.md
@@ -1,5 +1,11 @@
 ### HLAtools
 
+### Version 1.1.0
+
+- July 17, 2024
+- Added the GIANT() function for translating vectors and data frames of HLA allele names across IPD-IMGT/HLA Database release versions.
+- Updates to ReadMe and Vignette.
+
 ### Version 1.0.4
 
 - July 12, 2024